chore(magento): cached images
