+++ magento/loaders/cart.ts
@@ -1,14 +1,8 @@
-import {
-  default as extend,
-} from "../../website/loaders/extension.ts";
+import { default as extend } from "../../website/loaders/extension.ts";
 import { AppContext } from "../mod.ts";
-import {
-  getCartCookie,
-  getCartImages,
-  toCartItemsWithImages,
-} from "../utils/cart.ts";
-import { Cart as CartFromDeco, CartFromAPI } from "../utils/client/types.ts";
-import { ProductWithImages } from "../utils/clientGraphql/types.ts";
+import { handleCartImages } from "../utils/cache.ts";
+import { getCartCookie, toCartItemsWithImages } from "../utils/cart.ts";
+import { Cart as CartFromDeco } from "../utils/client/types.ts";
 import {
   BASE_CURRENCY_CODE,
   BASE_DISCOUNT_AMOUNT,
@@ -23,16 +17,9 @@ import {
 
 export type Cart = CartFromDeco;
 
-export const IMAGES_CACHE_NAME = "product-images-cache";
-
 interface Props {
   cartId?: string;
   disableExtensions?: boolean;
-}
-
-interface HandledImages {
-  cachedImages: ProductWithImages[];
-  nonCachedImagesSkus: string[];
 }
 
 /**
@@ -71,11 +58,11 @@ const loader = async (
         COUPON_CODE,
         BASE_CURRENCY_CODE,
       ].join(","),
-    }).then((c) => c.json()),
+    }).then((totalizers) => totalizers.json()),
     clientAdmin["GET /rest/:site/V1/carts/:cartId"]({
       cartId,
       site,
-    }).then((c) => c.json()),
+    }).then((items) => items.json()),
   ]);
 
   if (cart.items.length === 0) {
@@ -91,20 +78,12 @@ const loader = async (
     );
   }
 
-  const { cachedImages, nonCachedImagesSkus } = await handleCachedImages(
-    cart,
-    url,
-  );
-  const nonCachedImages = await handleNonCachedImages(
-    { skus: nonCachedImagesSkus, url, ctx },
-  );
+  const productsImages = await handleCartImages(cart, url, ctx);
 
   const cartWithImages = toCartItemsWithImages(
     cart,
     prices,
-    {
-      items: [...cachedImages, ...nonCachedImages],
-    },
+    productsImages,
     url.origin,
     site,
     countProductImageInCart,
@@ -118,63 +97,3 @@ const loader = async (
 };
 
 export default loader;
-
-const handleCachedImages = async (
-  { items }: CartFromAPI,
-  url: URL,
-): Promise<HandledImages> => {
-  const cachedImages: ProductWithImages[] = [];
-  const nonCachedImagesSkus: string[] = [];
-  const promises = items.map(async ({ sku }) => {
-    const image = await getCachedImages(sku, url);
-    if (image) {
-      cachedImages.push(image);
-      return;
-    }
-    nonCachedImagesSkus.push(sku);
-  });
-
-  await Promise.all(promises);
-
-  return {
-    cachedImages,
-    nonCachedImagesSkus,
-  };
-};
-
-const handleNonCachedImages = async (
-  { skus, url, ctx }: { skus?: string[]; url: URL; ctx: AppContext },
-) => {
-  if (!skus || skus.length === 0) {
-    return [];
-  }
-  //Get images from API
-  const { items } = (await getCartImages(skus, ctx)).products;
-
-  for (const i in items) {
-    putImageInCache(items[i], url);
-  }
-
-  return items;
-};
-
-const getCachedImages = async (
-  sku: string,
-  url: URL,
-): Promise<ProductWithImages | null> => {
-  const cacheKey = new URL(`images-${sku}`, url.origin);
-  const cache = await caches.open(IMAGES_CACHE_NAME);
-  const cachedResponse = await cache.match(cacheKey);
-
-  if (cachedResponse) {
-    return await cachedResponse.json() as ProductWithImages;
-  }
-  return null;
-};
-
-const putImageInCache = async (images: ProductWithImages, url: URL) => {
-  const cacheKey = new URL(`images-${images.sku}`, url.origin);
-  const cache = await caches.open(IMAGES_CACHE_NAME);
-  const responseToCache = new Response(JSON.stringify(images));
-  await cache.put(cacheKey, responseToCache);
-};
